… sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/TagsExtension.cs