* Updated `TagObjectsBenchmarks` to utilize the method that performs actual parsing
* Reduced `AzMonList` allocation size, 64 is a big number and does direct 1 KB allocation every time
* `AzMonList.Return` was not called, this is a easier way to release memory instead of waiting on GC
* Removed unused file

Old run using - `TagsExtension`

|                           Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|--------------------------------- |----------:|----------:|----------:|-------:|----------:|
|      Enumerate_TagObjects_NoItem |  73.19 ns |  2.072 ns |  5.777 ns | 0.0516 |     216 B |
|       Enumerate_TagObjects_PartB | 501.79 ns | 13.462 ns | 39.056 ns | 0.1431 |     600 B |
|       Enumerate_TagObjects_PartC | 948.24 ns | 24.602 ns | 70.190 ns | 0.2079 |     872 B |
| Enumerate_TagObjects_PartB_And_C | 439.67 ns | 14.627 ns | 41.016 ns | 0.1163 |     488 B |

New run using `AzMonList`

|                           Method |     Mean |    Error |   StdDev |   Median |  Gen 0 | Allocated |
|--------------------------------- |---------:|---------:|---------:|---------:|-------:|----------:|
|      Enumerate_TagObjects_NoItem | 174.5 ns |  3.48 ns | 10.04 ns | 175.0 ns |      - |         - |
|       Enumerate_TagObjects_PartB | 473.5 ns | 10.38 ns | 30.62 ns | 481.4 ns | 0.0095 |      40 B |
|       Enumerate_TagObjects_PartC | 759.5 ns | 15.55 ns | 45.86 ns | 773.8 ns | 0.0668 |     280 B |
| Enumerate_TagObjects_PartB_And_C | 454.3 ns | 11.03 ns | 32.52 ns | 467.0 ns | 0.0095 |      40 B |